+++ docs/.vitepress/theme/components/shipin.vue
@@ -1,0 +1,7 @@
+<template>
+    <div>
+      <video controls>
+        <source src="/lol.mp4" type="video/mp4">
+      </video>
+    </div>
+  </template>